Выполнил задание "Исключительный тип", провел рефакторинг кода
