+++ src/exceptions/NotValidStatusException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace taskforce\exceptions;
+
+class NotValidStatusException extends \Exception
+{
+
+}
