+++ src/models/actions/ApproveAction.php
@@ -1,10 +1,8 @@
 <?php
 
+namespace taskforce\models\actions;
 
-namespace taskforce\actions;
-
-
-use taskforce\Task;
+use taskforce\models\Task;
 
 class ApproveAction extends AbstractAction
 {
